Version 0.5.2.3-RC1

+++ mxflib/xmlparse.cpp
@@ -1,7 +1,7 @@
 /*! \file	xmlparse.cpp
  *	\brief	Interface to available SAX style XML parser
  *
- *	\version $Id: xmlparse.cpp,v 1.1 2005/03/24 15:19:50 terabrit Exp $
+ *	\version $Id: xmlparse.cpp,v 1.2 2005/09/26 08:35:59 matt-beard Exp $
  *
  */
 /*
@@ -71,9 +71,9 @@ bool mxflib::XMLParserParseFile(mxflib::XMLParserHandlerPtr Hand, void *UserData
 	do
 	{
 		const int BufferSize = 1024 * 64;
-		Uint8 *Buffer = (Uint8*)XML_GetBuffer(Parser, BufferSize);
+		UInt8 *Buffer = (UInt8*)XML_GetBuffer(Parser, BufferSize);
 
-		int Bytes = FileRead(InFile, Buffer, BufferSize);
+		int Bytes = (int)FileRead(InFile, Buffer, BufferSize);
 
 		if(FileEof(InFile)) Done = -1;
